Fix fixture color persistence compile errors and reuse viewer3d color logic
